+++ src/sdk/sdk.ts
@@ -61,9 +61,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "latest";
-    sdkVersion = "0.4.0";
-    genVersion = "2.237.3";
-    userAgent = "speakeasy-sdk/typescript 0.4.0 2.237.3 latest @lukehagar/discoursejs";
+    sdkVersion = "0.4.1";
+    genVersion = "2.253.0";
+    userAgent = "speakeasy-sdk/typescript 0.4.1 2.253.0 latest @lukehagar/discoursejs";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);