+++ RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v0.4.0] .
 ### Releases
 - [NPM v0.4.0] https://www.npmjs.com/package/@lukehagar/discoursejs/v/0.4.0 - .
+
+## 2024-02-13 01:08:58
+### Changes
+Based on:
+- OpenAPI Doc latest 
+- Speakeasy CLI 1.178.0 (2.253.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.4.1] .
+### Releases
+- [NPM v0.4.1] https://www.npmjs.com/package/@lukehagar/discoursejs/v/0.4.1 - .